Split AsAcceptOnlyInterface tests up per-test case
