+++ Source/Moq.Analyzers.Test/AsAcceptOnlyInterfaceAnalyzerTests.ShouldFailWhenUsingAsWithConcreteClass.verified.txt
@@ -1,0 +1,8 @@
+﻿Diagnostic 1
+	Id: Moq1300
+	Location: SourceFile(Test0.cs[642..652))
+	Highlight: OtherClass
+	Lines: mock.As<OtherClass>();
+	Severity: Error
+	Message: Mock.As() should take interfaces only
+
