--- Source/Moq.Analyzers.Test/AsAcceptOnlyInterfaceAnalyzerTests.ShouldPassIfGoodParameters.verified.txt
@@ -1,16 +1,0 @@
-﻿Diagnostic 1
-	Id: Moq1300
-	Location: SourceFile(Test0.cs[999..1014))
-	Highlight: BaseSampleClass
-	Lines: mock.As<BaseSampleClass>();
-	Severity: Error
-	Message: Mock.As() should take interfaces only
-
-Diagnostic 2
-	Id: Moq1300
-	Location: SourceFile(Test0.cs[1144..1154))
-	Highlight: OtherClass
-	Lines: mock.As<OtherClass>();
-	Severity: Error
-	Message: Mock.As() should take interfaces only
-
