+++ Source/Moq.Analyzers.Test/AsAcceptOnlyInterfaceAnalyzerTests.ShouldPassWhenUsingAsWithInterface.verified.txt
@@ -1,0 +1,1 @@
+﻿emptyString
