+++ Source/Moq.Analyzers.Test/Data/AsAcceptOnlyInterface.TestBadAsForNonAbstractClass.cs
@@ -1,0 +1,41 @@
+﻿using Moq;
+
+namespace AsAcceptOnlyInterface.TestBadAsForNonAbstractClass;
+
+public interface ISampleInterface
+{
+    int Calculate(int a, int b);
+
+    int TestProperty { get; set; }
+}
+
+public abstract class BaseSampleClass
+{
+    public int Calculate()
+    {
+        return 0;
+    }
+
+    public abstract int Calculate(int a, int b, int c);
+}
+
+public class SampleClass
+{
+
+    public virtual int Calculate(int a, int b) => 0;
+}
+
+public class OtherClass
+{
+
+    public virtual int Calculate() => 0;
+}
+
+internal class MyUnitTests
+{
+    private void TestBadAsForNonAbstractClass()
+    {
+        var mock = new Mock<BaseSampleClass>();
+        mock.As<OtherClass>();
+    }
+}
